feat(driver): remove transposition and refactor
Transposition has long be a source of confusion and errors and loss of speed.
Remove it in favor of using "manual tranposition" which leads to clear clean code.

BREAKING CHANGE: removed transposition, remove makeHTMLDriver entirely. Rename
makeDOMDriver to makeDomDriver.
