+++ src/DomSources/determineUseCapture.ts
@@ -1,0 +1,46 @@
+import { EventsFnOptions } from './DomSource';
+
+const eventTypesThatDontBubble = [
+  `blur`,
+  `canplay`,
+  `canplaythrough`,
+  `change`,
+  `durationchange`,
+  `emptied`,
+  `ended`,
+  `focus`,
+  `load`,
+  `loadeddata`,
+  `loadedmetadata`,
+  `mouseenter`,
+  `mouseleave`,
+  `pause`,
+  `play`,
+  `playing`,
+  `ratechange`,
+  `reset`,
+  `scroll`,
+  `seeked`,
+  `seeking`,
+  `stalled`,
+  `submit`,
+  `suspend`,
+  `timeupdate`,
+  `unload`,
+  `volumechange`,
+  `waiting`,
+]
+
+export function determineUseCapture (eventType: string, options: EventsFnOptions | undefined): boolean {
+  let result = false
+  if (!options) {
+    options = {}
+  }
+  if (typeof options.useCapture === `boolean`) {
+    result = options.useCapture as boolean
+  }
+  if (eventTypesThatDontBubble.indexOf(eventType) !== -1) {
+    result = true
+  }
+  return result
+}
