+++ src/DomSources/index.ts
@@ -1,0 +1,3 @@
+export * from './DomSource';
+export * from './ElementDomSource';
+export * from './MainDomSource';
