+++ src/index.ts
@@ -1,12 +1,11 @@
 // hyperscript
-export * from './interfaces'
+export * from 'snabbdom-ts';
 export * from './hyperscript/hyperscript'
 export * from './hyperscript/helpers'
-export * from './hyperscript/thunk'
-export * from './modules/index'
-export { DOMSource, EventsFnOptions } from './DOMSource'
-export { makeDOMDriver, DOMDriverOptions } from './makeDOMDriver'
-export { HTMLSource } from './HTMLSource'
-export { makeHTMLDriver, HTMLDriverOptions } from './makeHTMLDriver'
-export { mockDOMSource, MockConfig } from './mockDOMSource'
-export { MotorcycleDOMEvent } from './EventDelegator'
+import thunk from 'snabbdom-ts/thunk';
+export { thunk };
+export * from './modules/index';
+export * from './DomSources';
+export { makeDomDriver, DomDriverOptions } from './makeDomDriver'
+export { mockDomSource, MockConfig } from './mockDomSource'
+export { MotorcycleDomEvent } from './EventDelegator'
